Move the classes for writing to and reading from disk to the `storage` package, and the data with local user information to `preferences`.
