+++ client/src/main/kotlin/io/spine/examples/pingh/client/preferences/SnoozeTimeExts.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.examples.pingh.client
+package io.spine.examples.pingh.client.preferences
 
 import com.google.protobuf.Duration
 import com.google.protobuf.util.Durations
@@ -53,7 +53,7 @@ public val KClass<SnoozeTime>.supported: List<SnoozeTime>
 
 @Suppress("MagicNumber" /* The durations are specified using numbers. */)
 private val snoozeTimeDetails = mapOf(
-    SnoozeTime.THIRTY_MINUTES to SnoozeTimeDetails("30 mins", minutes(30)),
+    SnoozeTime.THIRTY_MINUTES to SnoozeTimeDetails("30 mins", minutes(1)),
     SnoozeTime.TWO_HOURS to SnoozeTimeDetails("2 hours", hours(2)),
     SnoozeTime.ONE_DAY to SnoozeTimeDetails("1 day", hours(24))
 )
