+++ client/src/main/kotlin/io/spine/examples/pingh/client/storage/FileStorage.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.examples.pingh.client
+package io.spine.examples.pingh.client.storage
 
 import com.google.common.annotations.VisibleForTesting
 import com.google.protobuf.Message
